cloud_functions/visits_to_bq/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.11.4
pandas==1.4.3